Rename charanames
