+++ tools/toir/toir/formats/dat/__init__.py
@@ -1,4 +1,4 @@
 from .extract import extract_dat
 from .recompile import recompile_dat
-from .charanames import read_chara_names
+from .packfield import read_chara_names
 from .datfile import DatFile
